# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [catalog]
      summary: List / search / filter products
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: categoryId
          in: query
          schema: { type: integer }
        - name: brandId
          in: query
          schema: { type: integer }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: color
          in: query
          schema: { type: string }
        - name: size
          in: query
          schema: { type: string }
        - name: minRating
          in: query
          schema: { type: number }
        - name: featured
          in: query
          schema: { type: boolean }
        - name: trending
          in: query
          schema: { type: boolean }
        - name: bestSeller
          in: query
          schema: { type: boolean }
        - name: newArrival
          in: query
          schema: { type: boolean }
        - name: onSale
          in: query
          schema: { type: boolean }
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc, rating, best_selling]
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 24 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResult"
    post:
      operationId: createProduct
      tags: [catalog]
      summary: Create product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [catalog]
      summary: Get a product by id or slug
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [catalog]
      summary: Update product (admin)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [catalog]
      summary: Delete product (admin)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [catalog]
      summary: Related products (same category)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductSummary"
  /products/{id}/reviews:
    get:
      operationId: listProductReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createProductReview
      tags: [reviews]
      summary: Create a review (auth required)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List current user's orders (auth required); admins see all
      parameters:
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Checkout - create order from cart (auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details (auth required)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/cancel:
    patch:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel order (auth required)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: integer }
  schemas:
...
    ProductSummary:
      type: object
      required:
        [
          id,
          name,
          slug,
          price,
          salePrice,
          images,
          rating,
          reviewCount,
          stock,
        ]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        price: { type: number }
        salePrice: { type: ["number", "null"] }
        images: { type: array, items: { type: string } }
        rating: { type: number }
        reviewCount: { type: integer }
        stock: { type: integer }
        featured: { type: boolean }
        trending: { type: boolean }
        bestSeller: { type: boolean }
        newArrival: { type: boolean }
    Product:
      type: object
      required:
        [
          id,
          name,
          slug,
          description,
          price,
          salePrice,
          stock,
          sku,
          images,
          colors,
          sizes,
          categoryId,
          brandId,
          tags,
          rating,
          reviewCount,
          featured,
          trending,
          bestSeller,
          newArrival,
          createdAt,
        ]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        richDescription: { type: ["string", "null"] }
        price: { type: number }
        salePrice: { type: ["number", "null"] }
        stock: { type: integer }
        sku: { type: string }
        barcode: { type: ["string", "null"] }
        weightGrams: { type: ["integer", "null"] }
        images: { type: array, items: { type: string } }
        colors: { type: array, items: { type: string } }
        sizes: { type: array, items: { type: string } }
        categoryId: { type: ["integer", "null"] }
        categoryName: { type: ["string", "null"] }
        brandId: { type: ["integer", "null"] }
        brandName: { type: ["string", "null"] }
        tags: { type: array, items: { type: string } }
        rating: { type: number }
        reviewCount: { type: integer }
        featured: { type: boolean }
        trending: { type: boolean }
        bestSeller: { type: boolean }
        newArrival: { type: boolean }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    ProductInput:
      type: object
      required: [name, slug, description, price, stock, sku, categoryId]
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        description: { type: string }
        richDescription: { type: string }
        price: { type: number, minimum: 0 }
        salePrice: { type: number, minimum: 0 }
        stock: { type: integer, minimum: 0 }
        sku: { type: string, minLength: 1 }
        barcode: { type: string }
        weightGrams: { type: integer }
        images: { type: array, items: { type: string } }
        colors: { type: array, items: { type: string } }
        sizes: { type: array, items: { type: string } }
        categoryId: { type: integer }
        brandId: { type: integer }
        tags: { type: array, items: { type: string } }
        featured: { type: boolean }
        trending: { type: boolean }
        bestSeller: { type: boolean }
        newArrival: { type: boolean }
        seoTitle: { type: string }
        seoDescription: { type: string }
    ProductUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        description: { type: string }
        richDescription: { type: string }
        price: { type: number, minimum: 0 }
        salePrice: { type: number, minimum: 0 }
        stock: { type: integer, minimum: 0 }
        sku: { type: string, minLength: 1 }
        barcode: { type: string }
        weightGrams: { type: integer }
        images: { type: array, items: { type: string } }
        colors: { type: array, items: { type: string } }
        sizes: { type: array, items: { type: string } }
        categoryId: { type: integer }
        brandId: { type: integer }
        tags: { type: array, items: { type: string } }
        featured: { type: boolean }
        trending: { type: boolean }
        bestSeller: { type: boolean }
        newArrival: { type: boolean }
        seoTitle: { type: string }
        seoDescription: { type: string }
    ProductListResult:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProductSummary"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    Review:
      type: object
      required:
        [id, productId, userName, rating, comment, verifiedPurchase, createdAt]
      properties:
        id: { type: integer }
        productId: { type: integer }
        userName: { type: string }
        rating: { type: integer }
        comment: { type: string }
        verifiedPurchase: { type: boolean }
        createdAt: { type: string, format: date-time }
    ReviewInput:
      type: object
      required: [rating, comment]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, minLength: 1 }
...
    OrderItem:
      type: object
      required:
        [
          id,
          productId,
          productName,
          productImage,
          unitPrice,
          quantity,
          color,
          size,
        ]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        productImage: { type: ["string", "null"] }
        unitPrice: { type: number }
        quantity: { type: integer }
        color: { type: ["string", "null"] }
        size: { type: ["string", "null"] }
    Order:
      type: object
      required:
        [
          id,
          orderNumber,
          status,
          items,
          subtotal,
          shippingCost,
          taxAmount,
          discountAmount,
          total,
          paymentMethod,
          shippingAddress,
          createdAt,
        ]
      properties:
        id: { type: integer }
        orderNumber: { type: string }
        status:
          type: string
          enum:
            [
              pending,
              processing,
              shipped,
              delivered,
              cancelled,
              refund_requested,
            ]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal: { type: number }
        shippingCost: { type: number }
        taxAmount: { type: number }
        discountAmount: { type: number }
        total: { type: number }
        paymentMethod: { type: string, enum: [cod, card, paypal] }
        couponCode: { type: ["string", "null"] }
        orderNotes: { type: ["string", "null"] }
        shippingAddress: { type: string }
        trackingStatus: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    OrderInput:
      type: object
      required: [addressId, paymentMethod]
      properties:
        addressId: { type: integer }
        paymentMethod: { type: string, enum: [cod, card, paypal] }
        couponCode: { type: string }
        orderNotes: { type: string }
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum:
            [
              pending,
              processing,
              shipped,
              delivered,
              cancelled,
              refund_requested,
            ]
...