# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderPaymentMethod } from './orderPaymentMethod';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  orderNumber: string;
  status: OrderStatus;
  items: OrderItem[];
  subtotal: number;
  shippingCost: number;
  taxAmount: number;
  discountAmount: number;
  total: number;
  paymentMethod: OrderPaymentMethod;
  /** @nullable */
  couponCode?: string | null;
  /** @nullable */
  orderNotes?: string | null;
  shippingAddress: string;
  /** @nullable */
  trackingStatus?: string | null;
  createdAt: Date;
}