# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  slug: string;
  description: string;
  /** @nullable */
  richDescription?: string | null;
  price: number;
  /** @nullable */
  salePrice: number | null;
  stock: number;
  sku: string;
  /** @nullable */
  barcode?: string | null;
  /** @nullable */
  weightGrams?: number | null;
  images: string[];
  colors: string[];
  sizes: string[];
  /** @nullable */
  categoryId: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  brandId: number | null;
  /** @nullable */
  brandName?: string | null;
  tags: string[];
  rating: number;
  reviewCount: number;
  featured: boolean;
  trending: boolean;
  bestSeller: boolean;
  newArrival: boolean;
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
  createdAt: Date;
}