# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  description: string;
  richDescription?: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  salePrice?: number;
  /** @minimum 0 */
  stock: number;
  /** @minLength 1 */
  sku: string;
  barcode?: string;
  weightGrams?: number;
  images?: string[];
  colors?: string[];
  sizes?: string[];
  categoryId: number;
  brandId?: number;
  tags?: string[];
  featured?: boolean;
  trending?: boolean;
  bestSeller?: boolean;
  newArrival?: boolean;
  seoTitle?: string;
  seoDescription?: string;
}